added in helpers.js file and removed findUserByEmail function from Express_Server and put it in helper function.
