+++ helpers.js
@@ -1,0 +1,14 @@
+
+
+const findUserByEmail = (email, database) => {
+  for (const userId in database) {
+    const user = database[userId];
+    if (user.email === email) {
+      return user;
+    }
+  }
+  return null;
+}
+
+
+module.exports = findUserByEmail
